Feat: Create useQuotesStore
Separate states and functions from useUserStore
